Added StemDetector
StemDetector receives a pre-processed pointcloud (already color-
filtered), and tries to align a cylidrical model to it. Two parameters
can be changed by the user: number of observations and model radius
limit. These parameters can be modified from the VisionConfig tab in
GUI.
StemDetector currently uses a simple method of computing the mean value
out of a specified number of observations. A simple threshold is applied
to remove the irrelevant observations.